chore(consensus): move conversions into respective modules
Moves the four `From<&BaseChainConfig>` impls from the top-level
`conversions.rs` into `hardfork.rs`, `params.rs`, `genesis.rs`, and
`rollup.rs`, then deletes `conversions.rs`.
