+++ crates/consensus/genesis/src/lib.rs
@@ -40,5 +40,3 @@ pub use rollup::{
     FJORD_MAX_SEQUENCER_DRIFT, GRANITE_CHANNEL_TIMEOUT, MAX_RLP_BYTES_PER_CHANNEL_BEDROCK,
     MAX_RLP_BYTES_PER_CHANNEL_FJORD, RollupConfig,
 };
-
-mod conversions;
